requirements.txt
ipython
pandas
DateTime
requests
python-dotenv
streamlit
flask
